% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_lp.R
\name{formula_lp}
\alias{formula_lp}
\title{explore formula of total points and linear predictors}
\usage{
formula_lp(nomogram, power, digits = 6)
}
\arguments{
\item{nomogram}{nomogram after nomogram command in rms package}

\item{power}{if missing, power will be choose automatically up to 100 based on all R2 equealling to 1}

\item{digits}{default is 6}
}
\value{
a global variable Formula_lp, the formula of total points and linear predictors
}
\description{
explore the total points formula to linear predictors and get best power.
}
\examples{
\donttest{
library(rms)
set.seed(2018)
n <-2019
age <- rnorm(n,60,20)
sex <- factor(sample(c('female','male'),n,TRUE))
sex <- as.numeric(sex)
weight <- sample(50:100,n,replace = TRUE)
time <- sample(50:800,n,replace = TRUE)
units(time)="day"
death <- sample(c(1,0,0),n,replace = TRUE)
df <- data.frame(time,death,age,sex,weight)
ddist <- datadist(df)
options(datadist='ddist')
f <- cph(formula(Surv(time,death)~sex+age+weight),data=df,
         x=TRUE,y=TRUE,surv=TRUE,time.inc=3)
surv <- Survival(f)
nomo <- nomogram(f,
                 lp=TRUE,
                 fun=list(function(x) surv(365,x),
                          function(x) surv(365*2,x)),
                 funlabel=c("1-Year Survival Prob",
                            "2-Year Survival Prob"))
library(nomogramFormula)
formula_lp(nomogram = nomo)
formula_lp(nomogram = nomo,power = 1)
formula_lp(nomogram = nomo,power = 2)
formula_lp(nomogram = nomo,power = 3,digits=6)
}

}
