% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomoFormu_probability_cat.R
\name{nomoFormu_probability_cat}
\alias{nomoFormu_probability_cat}
\title{extra and cat formula for probability formula}
\usage{
nomoFormu_probability_cat(nomogram, power)
}
\arguments{
\item{nomogram}{nomogram after nomogram command from rms package}

\item{power}{power, if missing, power will be choose automatically}
}
\value{
list and picture
}
\description{
explore the probability formula and get proper power.
}
\examples{
\donttest{
library(rms)
set.seed(2018)
n <-2019
age <- rnorm(n,60,20)
sex <- factor(sample(c('female','male'),n,TRUE))
sex <- as.numeric(sex)
weight <- sample(50:100,n,replace = TRUE)
time <- sample(50:800,n,replace = TRUE)
units(time)="day"
death <- sample(c(1,0,0),n,replace = TRUE)
df <- data.frame(time,death,age,sex,weight)
ddist <- datadist(df)
options(datadist='ddist')
f <- cph(formula(Surv(time,death)~sex+age+weight),data=df,
         x=TRUE,y=TRUE,surv=TRUE,time.inc=3)
surv <- Survival(f)
nomo <- nomogram(f,
                 lp=TRUE,
                 fun=list(function(x) surv(365,x),
                          function(x) surv(365*2,x)),
                 funlabel=c("1-Year Survival Prob",
                            "2-Year Survival Prob"))
library(nomogramFormula)
nomoFormu_probability_cat(nomogram = nomo,power = 1)
nomoFormu_probability_cat(nomogram = nomo,power = 2)
nomoFormu_probability_cat(nomogram = nomo,power = 3)
}
}
