% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_points.R
\name{total_points}
\alias{total_points}
\title{Caculate nomogram total points}
\usage{
total_points(data, lp, digits = 6)
}
\arguments{
\item{data}{data must be with no NA}

\item{lp}{linear predictors}

\item{digits}{default is 6}
}
\value{
total Points
}
\description{
Caculate nomogram total points
}
\examples{
\donttest{
library(rms)
set.seed(2018)
n <-2019
age <- rnorm(n,60,20)
sex <- factor(sample(c('female','male'),n,TRUE))
sex <- as.numeric(sex)
weight <- sample(50:100,n,replace = TRUE)
time <- sample(50:800,n,replace = TRUE)
units(time)="day"
death <- sample(c(1,0,0),n,replace = TRUE)
df <- data.frame(time,death,age,sex,weight)
ddist <- datadist(df)
options(datadist='ddist')
f <- cph(formula(Surv(time,death)~sex+age+weight),data=df,
         x=TRUE,y=TRUE,surv=TRUE,time.inc=3)
surv <- Survival(f)
nomo <- nomogram(f,
                 lp=TRUE,
                 fun=list(function(x) surv(365,x),
                          function(x) surv(365*2,x)),
                 funlabel=c("1-Year Survival Prob",
                            "2-Year Survival Prob"))
library(nomogramFormula)
#useing raw data to caculate total points
formula_points(nomo)
total_points(data=df)
#using linear predictors to caculate total points
f <- cph(formula(Surv(time,death)~sex+age+weight),data=df,
         linear.predictors=TRUE,
         x=TRUE,y=TRUE,surv=TRUE,time.inc=3)
surv <- Survival(f)
nomo <- nomogram(f,
                 lp=TRUE,
                 fun=list(function(x) surv(365,x),
                          function(x) surv(365*2,x)),
                 funlabel=c("1-Year Survival Prob",
                            "2-Year Survival Prob"))
formula_lp(nomo)
total_points(lp=f$linear.predictors)
}
}
